requirements.txt
transformers>=4.46.0
datasets>=3.0.0
accelerate>=1.0.0
peft>=0.13.0
trl>=0.12.0
torch>=2.3.0
sentencepiece>=0.2.0
wandb>=0.17.0
pyyaml>=6.0.0
scikit-learn>=1.5.0
tyro>=0.8.0

